+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v1.3.0 - Regularization
+---
+#### Network
+- Added dropout, with dropout configuration
+- Added L2 regularization, via l2 strength config
+- Added L1 regularization, via l1 strength config
+- Added max norm regularization, via the maxNorm threshold value config
+
+#### Bug Fixes
+- Fixed error value logged accumulating across epochs, instead of resetting to 0
+- Fixed epoch counter logged resetting instead of accumulating across training sessions 
+
 # v1.2.0 - Activation functions (Part 1)
 ---
 #### Network